Program-pipeline/requirements.txt
torch>=2.0.0
transformers>=4.30.0
dvc>=3.0.0
mlflow>=2.10.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pydantic>=2.0.0
apache-airflow>=2.8.0